requirements.txt
requests
urllib3
flask
Flask-BasicAuth
pyopenssl
dnacentersdk
python-dotenv
PyYAML
